refactor: move into state folder
