+++ src/data/constants-obsidian.ts
@@ -5,11 +5,6 @@
 export const CURRENT_PLUGIN_VERSION = "0.1.1";
 
 /**
- * The name of the view that will be registered with Obsidian
- */
-export const DASHBOARDS_VIEW = "dashboards";
-
-/**
  * The ID of the plugin
  * This should match the ID found in the manifest.json file
  */
