+++ src/react/container/container-content.tsx
@@ -1,5 +1,5 @@
 import { getMarkdownFromContainerContent } from "src/react/container/render-utils";
-import { Container } from "src/shared/types";
+import { Container } from "src/shared/state/types";
 import IconButton from "../icon-button/icon-button";
 import { useMountState } from "../mount-provider";
 import RenderMarkdown from "./render-markdown";
