+++ src/react/table/table.tsx
@@ -1,6 +1,6 @@
 import { css } from "@emotion/react";
 import Container from "../container/container";
-import { Container as ContainerType } from "src/shared/types";
+import { Container as ContainerType } from "src/shared/state/types";
 
 interface Props {
 	showBorders: boolean;
