+++ src/shared/constants.ts
@@ -1,2 +1,14 @@
+/**
+ * Global event id for when a user presses the ctrl or cmd key
+ */
 export const EVENT_CTRL_DOWN = "dashboards:ctrl-down";
+
+/**
+ * Global event id for when a user releases the ctrl or cmd key
+ */
 export const EVENT_CTRL_UP = "dashboards:ctrl-up";
+
+/**
+ * The name of the view that will be registered with Obsidian
+ */
+export const DASHBOARDS_VIEW = "dashboards";
